% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tagrates.R
\name{calc_tagrates}
\alias{calc_tagrates}
\title{Calculate tagrates from reshaped release data}
\usage{
calc_tagrates(release_data)
}
\arguments{
\item{release_data}{}
}
\description{
Calculate tagrates from reshaped release data
}
