% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luts.R
\name{download_luts}
\alias{download_luts}
\title{Download RMIS LUTs}
\usage{
download_luts(url = "ftp://ftp.rmpc.org/pub/data/", dir = NULL)
}
\arguments{
\item{url}{url to RMIS ftp serer}

\item{dir}{Directory to save RMIS lookup tables. If \code{NULL} (the default) puts luts in a RMIS_LUTs folder in working directory.}
}
\value{
Downloads lookup tables from RMIS into chosen dir
}
\description{
Downloads lookup tables from RMIS ftp server.
}
