% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_recoveries.R
\name{download_recoveries}
\alias{download_recoveries}
\title{Download CWT recoveries}
\usage{
download_recoveries(
  start_yr,
  end_yr,
  by_brood = TRUE,
  dir = RMIS.globals$rec_dir
)
}
\arguments{
\item{start_yr}{First year of recoveries}

\item{end_yr}{Last year of recoveries}

\item{by_brood}{\code{TRUE} (default) if start and end years are in terms of brood years}

\item{dir}{Directory to save downloaded files, if \code{NULL} creates Data/Recoveries/temp in working directory.}
}
\value{
\code{NULL}

Downloads .zip files of recovery data to local dir
}
\description{
Download CWT recoveries
}
