% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_releases.R
\name{download_releases}
\alias{download_releases}
\title{Download release data from RMIS}
\usage{
download_releases(file = "RL041_ALL_FULLSET.zip", dir = "RMIS")
}
\arguments{
\item{dir}{Directory to save downloaded file. If \code{NULL} (the default), saves in Data in working directory.}

\item{url}{the url to the full CWT release data set on the RMIS ftp server}
}
\value{
\code{NULL}

Saves file locally.
}
\description{
Download release data from RMIS
}
\examples{
download_releases()
}
