% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recoveries.R
\name{get_recoveries}
\alias{get_recoveries}
\title{Get recovery data}
\usage{
get_recoveries(
  first_by,
  last_by,
  ...,
  rec_dir = "RMIS/Recoveries",
  lut_dir = "RMIS/LUTs"
)
}
\arguments{
\item{first_by}{first brood year to retain in filter}

\item{last_by}{last brood year to retain in filter}

\item{...}{additional filter conditions passed in e.g., species_name=="Chinook"}

\item{rec_dir}{directory where recovery files are saved. Default = \code{RMIS/Recoveries}}

\item{lut_dir}{directory where lookup tables are saved. Default= \code{RMIS/LUTs}}
}
\value{

}
\description{
Get recovery data
}
