% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juv_tr_lut.R
\name{juv_tr_lut}
\alias{juv_tr_lut}
\title{Create a juvenile lookup table txt file}
\usage{
juv_tr_lut(first_by, last_by, ...)
}
\arguments{
\item{first_by}{first brood year to pull}

\item{last_by}{last brood year to pull}

\item{...}{filter conditions for decoded release data e.g., \code{species_name=="Chinook"}}
}
\value{

}
\description{
Create a juvenile lookup table txt file
}
