% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_list.R
\name{make_file_list}
\alias{make_file_list}
\title{Make a list of files to download}
\usage{
make_file_list(
  startYear,
  endYear,
  RMIS_url = "ftp://ftp.rmpc.org/pub/data/",
  dir = NULL
)
}
\arguments{
\item{startYear}{First year of desired CWT recovery data}

\item{endYear}{Last year of desired CWT recovery data}

\item{RMIS_url}{URL to RMIS ftp server}

\item{temp_dir}{File to hold downloaded csv's}
}
\value{
A list with download urls for each file and destination paths for each file
}
\description{
Creates a list of files to download for desired years of CWT recovery data.
}
