% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tag_info_lut.R
\name{make_tag_info_lut}
\alias{make_tag_info_lut}
\title{Make a tag info lut}
\usage{
make_tag_info_lut(first_by, last_by, ...)
}
\arguments{
\item{first_by}{first brood year of releases}

\item{last_by}{last brood year of releases}

\item{...}{other filter conditions passed in e.g., \code{species_name=="Chinook"}}
}
\value{
a table with release info for tag codes that meet brood year and other filter conditions
}
\description{
Make a tag info lut
}
