% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_releases.R
\name{read_releases}
\alias{read_releases}
\title{Read release data}
\usage{
read_releases(first_by = NULL, last_by = NULL, dir = "RMIS")
}
\arguments{
\item{first_by}{first brood year of data}

\item{last_by}{last brood year of data}

\item{dir}{Directory where release data were downloaded}
}
\value{
A data frame of the raw RMIS release data, filtered for brood years between first_by and last_by (inclusive)
}
\description{
Read release data
}
