% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases_for_tr.R
\name{releases_for_tr}
\alias{releases_for_tr}
\title{prep release data for tag rate calculations}
\usage{
releases_for_tr(release_data)
}
\arguments{
\item{release_data}{}
}
\value{
Table with a row for each tag_code_or_release_id, with 4 new columns for CWT and clip status: AdClipped_CWT, AdClipped_NoCWT, Unclipped_CWT, Unclipped_NoCWT and a composite key of species,hatchery,run,stock,release location for grouping when doing tagrate calcs
}
\description{
prep release data for tag rate calculations
}
