% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_release_filename.R
\name{set_release_filename}
\alias{set_release_filename}
\title{set the RMIS release file name}
\usage{
set_release_filename(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
set the RMIS release file name
}
