% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_recoveries.R
\name{tidy_recoveries}
\alias{tidy_recoveries}
\title{tidy up recovery data}
\usage{
tidy_recoveries(recoveries)
}
\arguments{
\item{recoveries}{A dataframe of recovery data from RMIS}
}
\value{
A dataframe with foreign keys decoded from LUTs
}
\description{
tidy up recovery data
}
