% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_luts.R
\name{download_luts}
\alias{download_luts}
\title{Download RMIS LUTs}
\usage{
download_luts(lut_dir = RMIS.globals$lut_dir)
}
\arguments{
\item{lut_dir}{Directory to save RMIS lookup tables. If \code{NULL} (the default) puts luts in a RMIS_LUTs folder in working directory.}

\item{url}{url to RMIS ftp serer}
}
\value{
\code{NULL}

Downloads lookup tables from RMIS into chosen dir
}
\description{
Downloads lookup tables from RMIS ftp server.
}
