% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_releases.R
\name{download_releases}
\alias{download_releases}
\title{Download release data from RMIS}
\usage{
download_releases(url = RMIS.globals$url, dir = RMIS.globals$rel_dir)
}
\arguments{
\item{url}{the url to the full CWT release data set on the RMIS ftp server}

\item{dir}{Directory to save downloaded file.  Defaults to \code{RMIS/Releases} in working directory.}
}
\value{
\code{NULL}

Saves file locally.
}
\description{
Download release data from RMIS
}
\examples{
download_releases()
}
