% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_combine_recoveries.R
\name{filter_and_combine_recoveries}
\alias{filter_and_combine_recoveries}
\title{Filter and combine recoveries}
\usage{
filter_and_combine_recoveries(
  first_by,
  last_by,
  ...,
  rec_dir = RMIS.globals$rec_dir,
  lut_dir = RMIS.globals$lut_dir
)
}
\arguments{
\item{first_by}{}

\item{last_by}{}

\item{...}{filter conditions for RMIS fields in release data}

\item{rec_dir}{Directory where recovery data were downloaded}
}
\description{
Filter and combine recoveries
}
