% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_combine_recoveries.R
\name{filter_and_combine_recoveries}
\alias{filter_and_combine_recoveries}
\title{Filter and combine recoveries}
\usage{
filter_and_combine_recoveries(
  first_by,
  last_by,
  rec_dir = "RMIS/Recoveries",
  lut_dir = "RMIS/LUTs",
  ...
)
}
\arguments{
\item{first_by}{}

\item{last_by}{}

\item{rec_dir}{Directory where recovery data were downloaded}

\item{...}{filter conditions for RMIS fields}
}
\value{

}
\description{
Filter and combine recoveries
}
