% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_data.R
\name{get_release_data}
\alias{get_release_data}
\title{Download dataframe of the RMIS release data}
\usage{
get_release_data(first_by = NULL, last_by = NULL, dir = NULL)
}
\arguments{
\item{first_by}{first brood year of releases}

\item{last_by}{last brood year of releases}

\item{dir}{directory where releases are held if \code{NULL} (the default) creates a Data folder in working directory.}
}
\value{
A dataframe of the RMIS release data with foreign keys decoded.
}
\description{
Download dataframe of the RMIS release data
}
