% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_releases.R
\name{group_releases}
\alias{group_releases}
\title{Add group ids to releases
groups multiple tag codes released from the same hatchery if released within 3 weeks and 4.5 grams avg wt.}
\usage{
group_releases(release_data)
}
\arguments{
\item{release_data}{}
}
\value{

}
\description{
Add group ids to releases
groups multiple tag codes released from the same hatchery if released within 3 weeks and 4.5 grams avg wt.
}
