% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases_for_tr.R
\name{releases_for_tr}
\alias{releases_for_tr}
\title{Tidy release data for tag rate calculations}
\usage{
releases_for_tr(release_data)
}
\arguments{
\item{release_data}{}
}
\value{
tidied data frame of release data by AD clip/CWT status
}
\description{
Tidy release data for tag rate calculations
}
